- Abicraft attribute system base done! Everything works - Attribute nodes, basic operations. - Added new test icons. - Small node ui changes

+++ Assets/ABICraft/Scripts/Monos/States/AttributeChangeTimedWrapper.cs
@@ -10,8 +10,8 @@ public class AttributeChangeTimedWrapper : AbicraftActionMono
     public AbicraftObject    obj;
     public AbicraftAttribute attribute;
 
-    public float setToValue;
-           float differenceToOriginal;
+    public int setToValue;
+           int differenceToOriginal;
 
 
     AbicraftAttribute.AbicraftObjectAttribute attributeObject;
